ModelService/requirements.txt
kafka-python==1.4.7
snappy==2.7